tests/wpt/web-platform-tests/html/semantics/embedded-content/the-canvas-element/toDataURL.zerowidth.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: toDataURL.zerowidth</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>toDataURL.zerowidth</h1>
<p class="desc">toDataURL on zero-size canvas returns 'data:,'</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="0"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("toDataURL on zero-size canvas returns 'data:,'");
_addTest(function(canvas, ctx) {

var data = canvas.toDataURL();
_assertSame(data, 'data:,', "data", "'data:,'");


});
</script>
